Boost exact title and NCN matches.
This Boosts results where the search term closely matches the title or ncn, so that the visually highlighted results in the PUI always (hopefully, I've chosen a boost value 'empirically' which seems to be sufficient) display at the top of the list of results.

I've tested this on all of Rose's top 10 judgments, and it works great on all but one, which is this https://caselaw.nationalarchives.gov.uk/ewca/civ/2022/1047 - it has an utterly pathological title with anonimisation and parentheticals which make it very difficult to match. However, on the upside, only two results are returned for the title query, (both of which have nearly identical titles), so it shouldn't impact user experience too much).