Add TDE files for stats extraction
This forms the basis of bulk data extraction for analysis; two new
Template Driven Extraction templates will populate `documents.summary`
and `documents.propertysummary` tables, which can be collectively
queried using an SQL `JOIN` to extract a table of key information about
documents.

These are a first pass, and will likely change as we get a better handle
on the stats we need to extract.

Fields which have `<invalid-values>` set to `ignore` and `<nullable>`
set to `true` will create rows with `NULL` values where a value cannot
be found. Rows where this isn't the case will be dropped, as they will
upset the `JOIN`.

| Field | Value | Ignore invalid? |
| --- | --- | --- |
| `uri` | `xdmp:node-uri(.)` | No |
| `ncn` | `sem:iri(//uk:cite//text())` | Yes |
| `name` | `sem:iri(//akn:FRBRWork/akn:FRBRname/@value)` | Yes |

| Field | Value | Ignore invalid? |
| --- | --- | --- |
| `uri` | `xdmp:node-uri(.)` | No |
| `doc_uri` | `dls:version/dls:document-uri` | Yes |
| `version_number` | `dls:version/dls:version-id` | Yes |
| `modified` | `prop:last-modified` | Yes |
| `published` | `published` | Yes |