Allow searching for multiple judges in a single query
This is to satisfy a specific request from the judicial working group - they've asked for the ability to do a limited 'boolean-OR' search specifically on the 'judge name' field, in order to be able to find all judgments by a judge who may have changed their name. This change transparently allows this - the field is separated on commas and combined in a boolean or-expression. Existing publicly announced behaviour is not affected, but those that know the trick can use commas!