Prioritise quoted phrases in search result snippets.
This depends on an upcoming PR on API client for two reasons.

First, we need to pass through the quoted phrases in the search from the client (Marklogic's regexp engine isn't up to the task, sadly)
And secondly, this transforms the snippet results after they are returned, and therefore we need to up the maximum number of snippets returned in order to make sure that we can pick out the ones we want. The changes in the API client ensure that only the first 4 snippets are displayed, to maintain the current behaviour.